% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_src_dist.r
\name{make_src_dist}
\alias{make_src_dist}
\title{make_src_dist}
\usage{
make_src_dist(path = ".", version, date = format(Sys.Date(), "\%B \%d \%Y"),
  verbose = TRUE, clobber = NULL)
}
\arguments{
\item{path}{The root directory of the release project (or any of its subfolders). The
file \code{.release_home} has to exist in some parent of \code{path}.}

\item{version}{A version string of the form \code{MAJOR.MINOR-PATCH}.}

\item{date}{The release date.}

\item{verbose}{Should the build system tell you what it's doing?}

\item{clobber}{An option meant to stop you from accidentally overwriting an existing source
release because you forgot to bump the version in your calling script.
\code{TRUE}: remove existing directory should it exist;
\code{FALSE}: pass through initial creation checks;
\code{NULL}: receive a warning which you can override by manually typing 'OK'
which then deletes the existing dir and starts over (typing anything else
halts with an error).}
}
\value{
The return is silent, but on successful exit, the source distribution will be
located in the path 'releases/pbdR_${version}/'.
}
\description{
Make a source pbdR distribution.
}
\details{
Downloads all packages as specified in the assets/spec.in file. CRAN packages
are the latest version. Github packages are the latest release (if none
exists, it'll break).
}
\examples{
\dontrun{
library(pbdRelease)

make_src_dist(version="1.0-0", verbose=TRUE)
}

}
