% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdRelease-package.r
\docType{package}
\name{pbdRelease-package}
\alias{pbdRelease-package}
\title{pbdRelease}
\description{
Tools for creating pbdR releases. If you aren't part of the pbdR core team,
this probably doesn't interest you.
}
\keyword{Package}
